requirements.txt
attrdict==2.0.1
matplotlib==3.7.4
ood-metrics==1.1.1
pandas==2.0.3
Pillow==10.1.0
PyYAML==6.0.1
tabulate==0.9.0
tqdm==4.66.1
